Hide warnings and add colors

+++ install/src/PonyInstall.hx
@@ -1,3 +1,5 @@
+import haxe.io.Eof;
+import sys.io.Process;
 import sys.FileSystem;
 
 /**
@@ -32,9 +34,34 @@ class PonyInstall extends BaseInstall {
 		}
 		log('Compile pony');
 		Utils.beginColor(90);
-		cmd('haxe', ['--cwd', Config.SRC, 'build.hxml']);
+		var newline: String = '\n';
+		var compiler: String = 'haxe';
+		var args: Array<String> = ['--cwd', Config.SRC, 'build.hxml'];
+		Sys.println(compiler + ' ' + args.join(' '));
+		var process: Process = new Process(compiler, args);
+		var r: Int = process.exitCode();
+		try {
+			while (true) Sys.println(process.stdout.readLine());
+		} catch (e: Eof) {}
+		try {
+			while (true) {
+				var line: String = process.stderr.readLine();
+				if (!checkWarning(line)) Sys.stderr().writeString(line + newline);
+			}
+		} catch (e: Eof) {}
+		if (r > 0) {
+			Utils.beginColor(31);
+			throw '$compiler error $r';
+		}
+		Utils.beginColor(32);
+		Sys.println('Compilation complete');
 		Utils.endColor();
 		FileSystem.deleteFile(Config.BIN + 'pony.n');
+	}
+
+	private function checkWarning(s: String): Bool {
+		if (s.indexOf(': Warning :') != -1) for (lib in Config.settings.hideWarnings) if (s.indexOf(lib) != -1) return true;
+		return false;
 	}
 
 }
